Combine startServer and stopServer to same file
