+++ client/package.json
@@ -37,8 +37,8 @@
     "babel-plugin-styled-components": "1.12.0",
     "browserslist": "4.16.6",
     "core-js": "3.14.0",
-    "i18next": "20.3.1",
-    "i18next-browser-languagedetector": "6.1.1",
+    "i18next": "20.3.2",
+    "i18next-browser-languagedetector": "6.1.2",
     "i18next-xhr-backend": "3.2.2",
     "react": "16.14.0",
     "react-beautiful-dnd": "13.1.0",
@@ -102,7 +102,7 @@
     "stylelint-config-styled-components": "0.1.1",
     "stylelint-no-unsupported-browser-features": "5.0.1",
     "stylelint-processor-styled-components": "1.10.0",
-    "webpack": "5.39.0",
+    "webpack": "5.39.1",
     "webpack-bundle-analyzer": "4.4.2",
     "webpack-cli": "4.7.2",
     "webpack-dev-server": "3.11.2",
