+++ package.json
@@ -56,7 +56,7 @@
     "ncp": "2.0.0",
     "rimraf": "3.0.2",
     "tsconfig-paths": "3.9.0",
-    "typescript": "4.3.2"
+    "typescript": "4.3.4"
   },
   "devDependencies": {
     "@types/eslint": "7.2.13",
@@ -69,7 +69,7 @@
     "@typescript-eslint/eslint-plugin": "4.27.0",
     "@typescript-eslint/parser": "4.27.0",
     "cross-env": "7.0.3",
-    "eslint": "7.28.0",
+    "eslint": "7.29.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-config-standard-with-typescript": "20.0.0",
     "eslint-import-resolver-babel-module": "5.3.1",
@@ -92,7 +92,7 @@
     "lint-staged": "11.0.0",
     "prettier": "2.3.1",
     "ts-jest": "27.0.3",
-    "type-fest": "1.2.0"
+    "type-fest": "1.2.1"
   },
   "engines": {
     "node": ">=12.x.x"
