+++ server/package.json
@@ -45,7 +45,7 @@
     "generate-serial-number": "0.0.3",
     "helmet": "4.6.0",
     "jsonwebtoken": "8.5.1",
-    "mongoose": "5.12.13",
+    "mongoose": "5.12.14",
     "morgan": "1.10.0",
     "munkres-js": "1.2.2",
     "node-schedule": "2.0.0",
@@ -60,9 +60,9 @@
     "@types/bcryptjs": "2.4.2",
     "@types/express": "4.17.12",
     "@types/faker": "5.5.6",
-    "@types/jsonwebtoken": "8.5.1",
+    "@types/jsonwebtoken": "8.5.2",
     "@types/morgan": "1.9.2",
-    "@types/node": "15.12.2",
+    "@types/node": "15.12.4",
     "@types/node-schedule": "1.3.1",
     "@types/request": "2.48.5",
     "@types/request-promise-native": "1.0.17",
