+++ server/src/index.ts
@@ -1,6 +1,5 @@
 import 'array-flat-polyfill';
-import { startServer } from 'server/utils/startServer';
-import { closeServer } from 'server/utils/closeServer';
+import { startServer, closeServer } from 'server/utils/server';
 import { logger } from 'server/utils/logger';
 import { autoUpdateGames, autoAssignPlayers } from 'server/utils/cron';
 import { config } from 'server/config';
